+++ src/farn/core/__init__.py
@@ -1,6 +1,13 @@
-from farn.core.parameter import Parameter as Parameter
+from farn.core.parameter import Parameter
 from farn.core.case import (
-    CaseStatus as CaseStatus,
-    Case as Case,
-    Cases as Cases,
+    CaseStatus,
+    Case,
+    Cases,
 )
+
+__all__ = [
+    "Case",
+    "CaseStatus",
+    "Cases",
+    "Parameter",
+]