+++ src/farn/sampling/__init__.py
@@ -1,0 +1,5 @@
+from farn.sampling.discrete import DiscreteSampling
+
+__all__ = [
+    "DiscreteSampling",
+]