+++ src/farn/core/case.py
@@ -18,11 +18,11 @@ from dictIO.utils.path import relative_path
 from numpy import ndarray
 from pandas import DataFrame, Series
 
-from farn.core import Parameter
+from farn.core.parameter import Parameter
 
-__ALL__ = [
-    "CaseStatus",
+__all__ = [
     "Case",
+    "CaseStatus",
     "Cases",
 ]