+++ CHANGELOG.md
@@ -8,6 +8,30 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 * -/-
 
 
+## [0.5.0] - 2026-04-09
+
+### BREAKING Changes
+* Moved module `farn.run.utils.threading` to  `farn.utils.threading`
+* Renamed and moved module `farn.run.subProcess` to `farn.utils.sub_process`
+* Renamed and moved module `farn.run.batchProcess` to `farn.batch.batch_processor`
+* Renamed module `farn.sampling.sampling` to `farn.sampling.discrete` (because the only class it actually contains is `DiscreteSampling`
+
+### Changed
+* Updated code base with latest changes in python_project_template v0.2.10
+* Python modules: Changed `__ALL__` to `__all__` (lowercase, PEP8 and PEP257 compliant).
+
+### Dependencies
+* .pre-commit-config.yaml: Updated rev of ruff-pre-commit to v0.15.9
+* Updated to dictIO>=0.4.4
+* Updated to numpy>=2.4
+* Updated to ospx>=0.3.5
+* Updated to pytest-cov>=7.1
+* Updated to ruff>=0.15.9
+* Updated to scipy>=1.17.1
+* Updated to sphinx-argparse-cli>=1.21.3
+* Updated to types-lxml>=2026.2
+
+
 ## [0.4.4] - 2026-03-17
 
 ### Solved